+++ src/components/inputs/Input.jsx
@@ -1,4 +1,4 @@
-import globalStyle from '../utils/globalStyle';
+import globalStyle from '../../utils/globalStyle';
 
 function Input({
   dataTestid,
