+++ src/components/todo/ModifyTodoForm.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import todoApi from '../../api/todo';
-import DefaultButton from '../DefaultButton';
-import Input from '../Input';
-import SubmitButton from '../SubmitButton';
+import DefaultButton from '../buttons/DefaultButton';
+import Input from '../inputs/Input';
+import SubmitButton from '../buttons/SubmitButton';
 
 function ModifyTodoForm({ todo, refreshHandler, closeHandler }) {
   const [modifyTodo, setMyTodo] = useState();
