+++ src/pages/SignUp/SignUp.jsx
@@ -1,10 +1,10 @@
 import { useNavigate } from 'react-router';
 import validationEmail from '../../utils/validationEmail';
 import validationPassword from '../../utils/validationPassword';
-import Input from '../../components/Input';
+import Input from '../../components/inputs/Input';
 import authApi from '../../api/auth';
 import useInputValidation from '../../hooks/useInputValidation';
-import SubmitButton from '../../components/SubmitButton';
+import SubmitButton from '../../components/buttons/SubmitButton';
 import ErrorMessage from '../../components/ErrorMessage';
 
 function SignUp() {
